Transposition Revamp + Housecleaning
Transposition Table agent was changed to prevent a predicted conflict
with being unable to have functions in reversi_board, this in turn
revealed another issue with my previous code where it was falsely
equating unequivalent board states.  This was resolved, but now the
Transposition Table agent is actively detrimental to efficiency in most
cases.

Also various housecleaning changes like moving to all_players